+++ src/it/ship/test/addtests/UploadNewsTest.java
@@ -1,10 +1,8 @@
-package it.ship.test;
+package it.ship.test.addtests;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.Statement;
@@ -18,18 +16,17 @@ import javax.servlet.http.HttpSession;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
+import it.ship.servlets.gestioneNews.UploadNews;
 import it.ship.servlets.gestioneSistema.DriverManagerConnectionPool;
-import it.ship.servlets.gestioneSistema.InserisciFeedback;
 
-class InserisciFeedbackTest {
+class UploadNewsTest {
 
 	HttpServletRequest requestMock = mock(HttpServletRequest.class);
 	HttpServletResponse responseMock = mock(HttpServletResponse.class);
 	HttpSession sessionMock = mock(HttpSession.class);
-	InserisciFeedback inserisciFeedback = mock(InserisciFeedback.class);
+	UploadNews uploadNews = mock(UploadNews.class);
 	RequestDispatcher dispatcherMock = mock(RequestDispatcher.class);
-	InserisciFeedback inserisciFeedbackTest = new InserisciFeedback();
+	UploadNews uploadNewsTest = new UploadNews();
 	
 	/*
 	 * Prima di ogni test simula la sessione dell'utente.
@@ -47,7 +44,7 @@ class InserisciFeedbackTest {
 		try {
 			Connection conn = DriverManagerConnectionPool.getConnection();
 			Statement stmtSelect = conn.createStatement();
-			stmtSelect.executeUpdate("DELETE FROM feedback WHERE Titolo='Parla';");
+			stmtSelect.executeUpdate("DELETE FROM news WHERE Autore='Mario Rossi';");
 			conn.close();
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -55,13 +52,14 @@ class InserisciFeedbackTest {
 	}
 	
 	@Test
-	void insertFeedback() throws ServletException, IOException {
-		when(requestMock.getParameter("testo")).thenReturn("Parla Parla");
-		when(requestMock.getParameter("id")).thenReturn("1");
-		when(requestMock.getParameter("user")).thenReturn("1");
-		when(requestMock.getParameter("titolo")).thenReturn("Parla");
-		inserisciFeedbackTest.doPost(requestMock, responseMock);
-		verify(responseMock).sendRedirect("/DettagliAlbum");
+	void uploadNews() throws ServletException, IOException {
+		when(requestMock.getParameter("newstitolo")).thenReturn("Nuova Canzone");
+		when(requestMock.getParameter("author")).thenReturn("Mario Rossi");
+		when(requestMock.getParameter("categoria")).thenReturn("Pop");
+		when(requestMock.getParameter("newscopertina")).thenReturn("Nuova Canzone");
+		when(requestMock.getParameter("newstext")).thenReturn("Canta");
+		uploadNewsTest.doPost(requestMock, responseMock);
+		verify(responseMock).sendRedirect("/pages/menager.jsp?cod=1");
 	}
 
 }
